requirements.txt
easyocr
opencv-python
torch
Pillow
numpy
logging
